+++ bin/git-docker.sh
@@ -33,6 +33,7 @@ if [ -f "${GIT_DOCKER_DIRECTORY}/git-docker-build.sh" ];    then  source "${GIT_
 if [ -f "${GIT_DOCKER_DIRECTORY}/git-docker-info.sh" ];     then  source "${GIT_DOCKER_DIRECTORY}/git-docker-info.sh";      fi
 if [ -f "${GIT_DOCKER_DIRECTORY}/git-docker-shell.sh" ];    then  source "${GIT_DOCKER_DIRECTORY}/git-docker-shell.sh";     fi
 if [ -f "${GIT_DOCKER_DIRECTORY}/git-docker-create.sh" ];   then  source "${GIT_DOCKER_DIRECTORY}/git-docker-create.sh";      fi
+if [ -f "${GIT_DOCKER_DIRECTORY}/git-docker-clean.sh" ];     then  source "${GIT_DOCKER_DIRECTORY}/git-docker-clean.sh";      fi
 if [ -f "${GIT_DOCKER_DIRECTORY}/git-docker-list.sh" ];     then  source "${GIT_DOCKER_DIRECTORY}/git-docker-list.sh";      fi
 
 ## Route Actions.
@@ -45,6 +46,7 @@ if [[ ${GIT_DOCKER_ACTION} == "reload" ]];    then GitDockerReload    $2 $3;    
 if [[ ${GIT_DOCKER_ACTION} == "build" ]];     then GitDockerBuild     $2 $3;    fi;
 if [[ ${GIT_DOCKER_ACTION} == "shell" ]];     then GitDockerShell     $2 $3;    fi;
 if [[ ${GIT_DOCKER_ACTION} == "list" ]];      then GitDockerList      $2 $3;    fi;
+if [[ ${GIT_DOCKER_ACTION} == "clean" ]];     then GitDockerClean     $2 $3;    fi;
 
 if [[ ${GIT_DOCKER_ACTION} == "" ]]; then
   echo "Please specify a command such as [info], [start] or [stop]. Showing [git docker info]."