+++ changes.md
@@ -1,3 +1,8 @@
+#### 1.0.2
+* Added [git docker clean] command to remove all Git Config settings.
+* Added automatic removal of .git from remote URL.
+* The PID of container is recorded to docker.meta.pid.
+
 #### 1.0.1
 * Added container port storage to Git config.
 * Increased default memory limit to 6g from 3g.